+++ CHANGELOG.md
@@ -146,3 +146,8 @@
 
 **Config**
 - travelersbackpack (M)
+
+## Modpack update — 2026-07-26
+
+**Client data**
+- azc-riding-client.zip (M)